+++ .github/workflows/ci_test-full.yml
@@ -60,7 +60,7 @@ jobs:
     - name: Set min. dependencies
       if: matrix.requires == 'oldest'
       run: |
-        python .github/set-oldest-versions.py
+        python .github/assistant.py set-oldest-versions
 
     - run: echo "::set-output name=period::$(python -c 'import time ; days = time.time() / 60 / 60 / 24 ; print(int(days / 7))' 2>&1)"
       if: matrix.requires == 'latest'
